backend/requirements.txt

flask==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
flask-socketio==5.3.3
mysql-connector-python==8.0.32
opencv-python==4.7.0.72
requests==2.28.2
werkzeug==2.2.3
gunicorn==20.1.0
python-dotenv==1.0.0
</lov-add-dependency>
